session: fix pacer bucket update cast
Make sure comparison is done between two i64 values.

Type: fix

+++ src/vnet/session/transport.c
@@ -665,7 +665,8 @@ spacer_max_burst (spacer_t * pacer, clib_us_time_t time_now)
   if ((inc = (f32) n_periods * pacer->tokens_per_period) > 10)
     {
       pacer->last_update = time_now;
-      pacer->bucket = clib_min (pacer->bucket + inc, pacer->max_burst);
+      pacer->bucket =
+	clib_min ((i64) (pacer->bucket + inc), (i64) pacer->max_burst);
     }
 
   return pacer->bucket > 0 ? pacer->max_burst : 0;
